(svn r4470) -Fix: FS#97 — Possible bug in Win64 versions (by michi_cc)   Doesn't fix any known bug, but the code is now bit cleaner. The proper result of subtraction of two pointers is ptrdiff_t.

+++ saveload.c
@@ -722,7 +722,7 @@ void SlObject(void *object, const SaveLoad *sld)
 	}
 
 	for (; sld->cmd != SL_END; sld++) {
-		void *ptr = (byte*)object + (unsigned long)sld->address;
+		void *ptr = (byte*)object + (ptrdiff_t)sld->address;
 		SlObjectMember(ptr, sld);
 	}
 }
